tickets/requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
asyncio==3.4.3
Django==3.2.11
django-grpc==1.0.16
djangorestframework==3.13.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.43.0
grpcio-tools==1.43.0
promise==2.3
protobuf==3.19.3
psycopg2-binary==2.9.3
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==4.0.1
